requirements.txt
streamlit
weaviate-client
python-dotenv
pandas
openai 
faiss-cpu